requirements.txt
# Ques Backend - Merged Requirements
# Combines dependencies from both backend_p12 and backend_p34
# Install: pip install -r requirements.txt

# Core FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data validation and configuration  
pydantic[email]>=2.9.0
email-validator>=2.0.0

# HTTP requests for external APIs
requests==2.31.0
aiohttp==3.8.6
httpx==0.25.0

# Environment variables
python-dotenv==1.0.0

# PostgreSQL database
psycopg2-binary>=2.9.7
sqlalchemy>=2.0.25
alembic>=1.12.0

# Authentication and Security
pyjwt>=2.8.0
bcrypt>=4.0.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# File uploads and multipart
python-multipart>=0.0.6

# Database migrations
alembic==1.13.0

# JWT and Authentication
python-jose[cryptography]==3.3.0
pyjwt>=2.8.0

# Password hashing and security
passlib[bcrypt]==1.7.4
cryptography>=41.0.0

# Email validation
email-validator==2.0.0

# Enhanced security and monitoring
structlog>=23.0.0

# Redis for caching and sessions
redis>=5.0.0
hiredis>=2.3.0

# Vector embeddings and AI
sentence-transformers>=2.7.0
huggingface_hub>=0.19.0
numpy>=1.24.0
torch>=2.0.0
transformers>=4.30.0

# Tencent Cloud Services
tcvectordb
tencentcloud-sdk-python>=3.0.1000

# DeepSeek API for tag extraction
openai>=1.0.0  # Compatible API client

# JSON parsing and text processing
jsonschema>=4.0.0

# WebSocket support for real-time features
websockets>=11.0.0

# Rate limiting
slowapi>=0.1.9

# Image processing (for profile pictures and project images)
Pillow>=10.0.0

# File type detection
python-magic>=0.4.27

# Background tasks
celery>=5.3.0

# Additional utilities
python-multipart>=0.0.6  # For file uploads
httpx>=0.25.0  # Modern HTTP client
pytest>=7.4.0  # Testing framework
pytest-asyncio>=0.21.0  # Async testing

# Development dependencies (optional)
black>=23.0.0  # Code formatting
isort>=5.12.0  # Import sorting
flake8>=6.0.0  # Linting
mypy>=1.5.0  # Type checking
